refactor: DTO 패키지 controller로 이동
