+++ src/main/java/roomescape/controller/dto/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller.dto;
 
 import java.time.LocalTime;
 import roomescape.domain.ReservationTime;
