+++ src/test/java/roomescape/e2e/ReservationTimeApiE2ETest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.dto.ReservationTimeRequest;
+import roomescape.controller.dto.ReservationTimeRequest;
 
 import static org.hamcrest.CoreMatchers.is;
